+++ startsmart/campaigns/admin.py
@@ -1,6 +1,6 @@
 from django.contrib import admin
-from .models import Profile
+from .models import Campaign
 
 # Register your models here.
 
-admin.site.register(Profile)
+admin.site.register(Campaign)
