+++ startsmart/campaigns/apps.py
@@ -1,0 +1,7 @@
+from django.apps import AppConfig
+
+
+
+print("YAAAA")
+class CampaignsConfig(AppConfig):
+    name = 'startsmart.campaigns'
