+++ startsmart/templates/users/main.html
@@ -23,9 +23,6 @@
                 <a class="nav-link" href="{% url 'profile' request.user.profile.id %}">Profile</a>
               </li>
               <li class="nav-item">
-                  <a class="nav-link" href="{% url 'create-campaign' %}">Create Campaign</a>
-              </li>
-              <li class="nav-item">
                   <a class="nav-link" href="{% url 'logout' %}">Logout</a>
               </li>
             </ul>
