+++ startsmart/users/apps.py
@@ -1,5 +1,7 @@
 from django.apps import AppConfig
 
+
+print("SUUUP")
 class UsersConfig(AppConfig):
     name = 'startsmart.users'
