+++ startsmart/users/forms.py
@@ -1,6 +1,6 @@
 from django.forms import ModelForm
 from django import forms
-from .models import Profile, Campaign
+from .models import Profile
 
 from django.contrib.auth.forms import UserCreationForm
 from django.contrib.auth.models import User
@@ -35,14 +35,3 @@ class UpdateProfileForm(forms.ModelForm):
 
 ##Member choice members: (Free, Pro, Elite) or (Customer, Pro, Creator)
 ##MEMBER_CHOICES = [(), (), ()]
-
-class CampaignForm(forms.ModelForm):
-
-    class Meta:
-        model = Campaign
-        fields = '__all__'
-        
-        widgets = {
-                'user': forms.TextInput(attrs={'value': '', 'id': 'element', 'type': 'hidden'})
-            }
-
